refactor(cli): extract lifecycle and turn command owners
